requirements.txt
# Core Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database Dependencies
pymongo>=4.6.0
neo4j>=5.15.0
influxdb-client>=1.39.0
redis>=5.0.1

# AI/ML Dependencies (Cutting-Edge Neural Networks)
tensorflow>=2.15.0
torch>=2.1.1
torchvision>=0.16.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Advanced Neural Network Dependencies
keras>=2.15.0
pytorch-lightning>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
tokenizers>=0.15.0

# Deep Learning Optimization
optax>=0.1.7
jax>=0.4.20
flax>=0.7.5

# Neural Network Visualization
tensorboard>=2.15.0
wandb>=0.16.0
netron>=7.2.0

# Advanced Data Processing for AI
dask>=2023.12.0
polars>=0.20.0
pyarrow>=14.0.0

# Behavioral Analysis
opencv-python>=4.8.0
mediapipe>=0.10.8
face-recognition>=1.3.0

# Time Series Analysis for Behavioral Patterns
statsmodels>=0.14.0
prophet>=1.1.4
tslearn>=0.6.2

# NLP Dependencies (Optional)
# transformers>=4.36.0
# nltk>=3.8.1
# spacy>=3.7.0

# Network and Security
scapy>=2.5.0
paramiko>=3.4.0
cryptography>=41.0.0
pyopenssl>=23.3.0

# Authentication and JWT
PyJWT>=2.8.0
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Virtualization and Containerization
docker==6.1.3
kubernetes==28.1.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Web Scraping and HTTP
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
jinja2==3.1.2

# Data Processing
celery==5.3.4
apache-airflow==2.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# API Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
rich==13.7.0

# Quantum Security Enhancements
pynacl==1.5.0
argon2-cffi==23.1.0
pycryptodome==3.19.0

# Enterprise Dashboard Dependencies
bokeh==3.3.2
plotly==5.17.0
dash==2.16.1

# Advanced Threat Intelligence
yara-python==4.3.1
pefile==2023.2.7
python-magic==0.4.27

# Real-time Analytics
elasticsearch==8.11.0
kafka-python==2.0.2

# Advanced AI/ML
xgboost==2.0.2
lightgbm==4.1.0
optuna==3.4.0

# Enterprise Reporting
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9

# Advanced Monitoring
grafana-api==1.0.3
prometheus-api-client==0.5.3

# WebSocket Support
websockets==12.0

# Additional Security
bandit==1.7.5
safety==2.3.5

# Cutting-Edge Autonomous Response System
asyncio-mqtt>=0.16.0
paho-mqtt>=1.6.1
celery[redis]>=5.3.4
kombu>=5.3.4

# Zero-Trust Architecture
python-ldap>=3.4.3
pysaml2>=7.4.2
oauthlib>=3.2.2
authlib>=1.2.1

# Advanced Behavioral Biometrics
keystroke-dynamics>=0.1.0
mouse-dynamics>=0.1.0
gait-analysis>=0.1.0

# Graph Neural Networks
torch-geometric>=2.4.0
dgl>=1.1.3
networkx>=3.2.1
igraph>=0.11.3

# Quantum Computing Simulation
qiskit>=0.45.0
cirq>=1.3.0
pennylane>=0.33.0

# Advanced Cryptography
pqcrypto>=0.1.0
liboqs-python>=0.8.0
post-quantum-cryptography>=0.1.0

# Real-time Stream Processing
apache-beam>=2.52.0
flink>=1.18.0
storm>=2.5.0

# Advanced Monitoring and Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
jaeger-client>=4.8.0
zipkin>=0.20.2

# High-Performance Computing
numba>=0.58.0
cupy>=12.3.0
rapids-cudf>=23.12.0

# Advanced Visualization
plotly-dash>=2.16.1
streamlit>=1.28.0
gradio>=4.8.0

# Enterprise Integration
apache-kafka>=2.0.2
rabbitmq>=0.2.0
zeromq>=25.1.1

# Advanced Analytics
apache-spark>=3.5.0
dask-ml>=2023.3.24
ray>=2.8.0

# Cutting-Edge Security Tools
volatility3>=2.5.0
yara>=4.3.1
sigma>=0.21.0

# Performance Optimization
cython>=3.0.6
pypy>=7.3.13
numexpr>=2.8.7 